generate-database: Remove obsolete commands stmt and method
These commands are replaced by generate
